Add files via upload
Virtual Garden is a user-friendly website that helps users to find medicinal plants through selection of body parts or the diseases. This provides users with the name of the plant, recipe and the benefits of its usage. An initiative to spread awareness among the people regarding medicinal plants through Technology