Create README.md
Virtual Garden was designed to implement the problem statement taken from the organization "Ministry of Ayush" from Smart India Hackathon 2024.
It aims to spread awareness on the medicinal plants through Technology. For those who are willing to use medicinal plants for healthy being or upon any diagnosis, this website stands as a guide to know more about the plants that works as medicine for the selected body part or the disease by the user. The website provides them with the names of medicinal plants, it's recipe and the benefits of it's usage.